Move interfaces to types.ts file
